requirements.txt
Flask
python-dotenv
flask-sqlalchemy
flask-cors
sqlalchemy-seeder
setuptools
requests
